processor/requirements.txt
boto3==1.34.122
demucs==4.1.0a2
torchaudio==2.3.1
soundfile==0.12.1
python-dotenv==1.0.1